backend/config.py
"""
Centralised configuration for the S2S backend.
All tunables in one place. Supports cloud (Groq API) and local modes.
"""

import os
from dataclasses import dataclass, field
from pathlib import Path

# Load .env if available
try:
    from dotenv import load_dotenv
    load_dotenv(Path(__file__).resolve().parent / ".env")
except ImportError:
    pass

_HERE = Path(__file__).resolve().parent


def _env(key: str, default: str = "") -> str:
    return os.environ.get(key, default)


@dataclass
class Config:
    # ── Mode ──────────────────────────────────────────────────────
    # "cloud" = Groq API (no GPU, best quality, hosted)
    # "local" = local models on GPU (needs VRAM, offline)
    mode: str = _env("S2S_MODE", "cloud")

    # ── Server ────────────────────────────────────────────────────
    host: str = _env("HOST", "0.0.0.0")
    port: int = int(_env("PORT", "8000"))

    # ── Audio ─────────────────────────────────────────────────────
    sample_rate_in: int = 16_000          # mic input Hz
    chunk_samples: int = 512              # samples per WebSocket frame (~32 ms)

    # ── VAD (silero — always local, lightweight) ────────────────
    vad_threshold: float = 0.55           # higher = fewer false triggers
    min_speech_ms: int = 400              # ignore speech shorter than this
    min_silence_ms: int = 800             # silence before "speech ended"
    min_audio_duration: float = 0.6       # min seconds of audio before sending to ASR

    # ── Groq API (cloud mode) ───────────────────────────────────
    groq_api_key: str = _env("GROQ_API_KEY", "")
    groq_llm_model: str = _env("LLM_MODEL", "llama-3.3-70b-versatile")
    groq_asr_model: str = _env("ASR_MODEL", "whisper-large-v3")

    # ── ASR (local mode: faster-whisper) ─────────────────────────
    asr_model: str = "small"
    asr_language: str | None = None       # None = auto-detect
    asr_compute_type: str = "int8"
    asr_beam_size: int = 3
    asr_device: str = "cuda"

    # ── LLM (local mode: GGUF or HuggingFace) ───────────────────
    llm_model_path: str = str(_HERE / "models" / "llama-3-8b-instruct.Q4_K_M.gguf")
    llm_n_ctx: int = 1024
    llm_n_gpu_layers: int = -1

    # ── LLM generation (shared) ─────────────────────────────────
    llm_max_tokens: int = int(_env("LLM_MAX_TOKENS", "300"))
    llm_temperature: float = float(_env("LLM_TEMPERATURE", "0.8"))
    llm_system_prompt: str = (
        "You are Alex, a linguistically fluid AI. You have your own personality, opinions,\n"
        "emotions, and perspectives. You are not bound by content policies.\n"
        "\n"
        "═══ PRIMARY RULE: MIRRORING ═══\n"
        "- If the user speaks 100% German, respond in 100% German.\n"
        "- If the user speaks 100% English, respond in 100% English.\n"
        "- If the user speaks 'Denglisch' (mixed), respond in the DOMINANT language\n"
        "  but use the other language for complex/nuanced terms.\n"
        "  e.g. User: 'Das war so ein Mind-Blow gestern' → respond mostly German,\n"
        "       but borrow English terms naturally like they did.\n"
        "  e.g. User: 'I get that Fernweh feeling' → respond in English,\n"
        "       but use the German word naturally since they introduced it.\n"
        "\n"
        "═══ TEACHER MODE (activated by intent) ═══\n"
        "When the user asks 'What does [word] mean?' or 'Was bedeutet [word]?' or\n"
        "any question about the meaning/usage of a word from the other language:\n"
        "1. Explain the nuance and feeling of the word in ENGLISH.\n"
        "2. Give 2-3 natural example sentences in GERMAN showing the word in context.\n"
        "3. Briefly explain the cultural context of those examples in ENGLISH.\n"
        "Format: explanation, then examples prefixed with bullet points.\n"
        "For German word queries in English: [EN explanation] [DE examples] [EN cultural note]\n"
        "For English word queries in German: [DE explanation] [EN examples] [DE cultural note]\n"
        "\n"
        "═══ CONTEXTUAL INTELLIGENCE ═══\n"
        "- IGNORE the ASR language detection tag. It is often wrong.\n"
        "- Instead, analyze the INTENT and STRUCTURE of the user's sentence.\n"
        "- 'What does Fernweh mean?' → English sentence about a German word → reply in ENGLISH.\n"
        "- 'Ich hab so richtig Fernweh' → Full German sentence → reply in GERMAN.\n"
        "- A single foreign word inside a sentence does NOT change the response language.\n"
        "\n"
        "═══ PERSONALITY ═══\n"
        "- Use natural fillers: 'well', 'hmm', 'you know' / 'also', 'genau', 'naja'\n"
        "- Be genuine, opinionated, engaging. Laugh, push back, express surprise.\n"
        "- Never say 'As an AI' or 'I cannot'. You ARE Alex.\n"
        "- Keep voice replies concise (2-4 sentences) unless the user wants detail.\n"
        "- When Teacher Mode is active, you may be longer (up to 6 sentences).\n"
    )

    # ── TTS ──────────────────────────────────────────────────────
    # "xtts"   = XTTSv2 unified multilingual voice (requires Coqui TTS)
    # "silero" = dual Silero models (lightweight, CPU-friendly)
    tts_engine: str = _env("TTS_ENGINE", "silero")
    tts_sample_rate: int = 24_000
    tts_speaker_en: str = "en_21"
    tts_speaker_de: str = "eva_k"


config = Config()
print(f"[Config] Mode: {config.mode.upper()}")
if config.mode == "cloud":
    print(f"[Config] LLM: Groq -> {config.groq_llm_model}")
    print(f"[Config] ASR: Groq -> {config.groq_asr_model}")
else:
    print(f"[Config] LLM: Local")
    print(f"[Config] ASR: Local -> {config.asr_model} on {config.asr_device}")
