requirements.txt
# ══════════════════════════════════════════════════════════
# Bilingual Speech-to-Speech AI — Dependencies
# ══════════════════════════════════════════════════════════

# ── Core ─────────────────────────────────────────────────
torch>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0               # .env file support

# ── Server ───────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# ── Cloud APIs (cloud mode) ─────────────────────────────
groq>=0.4.0                        # Groq API: LLM + Whisper ASR

# ── ASR (local mode) ────────────────────────────────────
faster-whisper>=1.0.0              # CTranslate2 backend (fast, CPU/GPU)
openai-whisper>=20231117           # Fallback ASR

# ── LLM (local mode) ────────────────────────────────────
transformers>=4.36.0               # Qwen2.5-1.5B-Instruct
accelerate>=0.25.0
sentencepiece>=0.1.99
# llama-cpp-python>=0.2.56         # Optional: GGUF models (install manually for GPU)

# ── TTS / VAD (always local — lightweight) ──────────────
# Silero TTS + VAD loaded via torch.hub or local .pt/.jit files
